Updated .versions to match package version 2.0.0-rc.1
When including 2.0.0-rc.1 in a meteor app, a version conflict arises
Conflict: Constraint jagi:astronomy@2.0.0-rc.1 is not satisfied by jagi:astronomy 1.2.10.
Constraints on package "jagi:astronomy":

Updating the package .versions to remove the packages no longer required and updating
the package version for astronomy to 2.0.0-rc.1 allows the package to be installed in
meteor apps without conflict.